requirements-dev.txt
jsonschema>=4.21.0
openapi-spec-validator>=0.7.1
PyYAML>=6.0.1
